requirements.txt
python-telegram-bot>=20,<23
web3>=6,<8
asyncpg>=0.27,<1
python-dotenv>=1,<2
